Support minimal shared crate for client / server
