+++ Cargo.toml
@@ -11,24 +11,26 @@ license-file = "LICENSE"
 keywords = ["akd", "key transparency"]
 
 [workspace.dependencies]
+akd = { version = "=0.11.0", features = ["whatsapp_v1", "public_auditing"] }
+akd_core = { version = "=0.11.0" }
 akd_watch_auditor = { path = "crates/auditor" }
+akd_watch_common = { path = "crates/common" }
+akd_watch_protocol = { path = "crates/protocol" }
 akd_watch_watcher = { path = "crates/watcher" }
 akd_watch_web = { path = "crates/web" }
-akd_watch_common = { path = "crates/common" }
 anyhow = "=1.0.98"
 chrono = { version = "=0.4.41", features = ["serde"] }
 config = "=0.14.1"
+ed25519-dalek = "=2.2.0"
+hex = "=0.4.3"
+mockito = "=1.6.1"
+rand = "=0.9.3"
+serde = { version = "=1.0.219", features = ["derive"] }
+serde_json = { version = "=1.0.142" }
+thiserror = "=2.0.12"
+tokio = { version = "=1.46.1", features = ["full"] }
 tracing = "=0.1.41"
 tracing-subscriber = "=0.3.20"
-thiserror = "=2.0.12"
-hex = "=0.4.3"
-tokio = { version = "=1.46.1", features = ["full"] }
-akd = { version = "=0.11.0", features = ["whatsapp_v1", "public_auditing"] }
-akd_core = { version = "=0.11.0" }
-serde_json = { version = "=1.0.142" }
-serde = { version = "=1.0.219", features = ["derive"] }
-ed25519-dalek = "=2.2.0"
-rand = "=0.9.3"
 
 [workspace.lints.clippy]
 unused_async = "deny"
