+++ crates/protocol/Cargo.toml
@@ -1,33 +1,31 @@
 [package]
-name = "akd_watch_common"
+name = "akd_watch_protocol"
 
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
 keywords.workspace = true
+build = "src/build.rs"
 
 [features]
-testing = ["akd_watch_protocol/testing"]
+testing = []
 
 [dependencies]
-akd = { workspace = true }
-akd_watch_protocol = { workspace = true }
 bincode = { version = "=2.0.1", features = ["serde"] }
 chrono = { workspace = true }
-config = { workspace = true }
 ed25519-dalek = { workspace = true, features = ["serde"] }
 hex = { workspace = true }
-quick-xml = "=0.38.0"
+prost = "=0.13.5"
 rand = { workspace = true }
-reqwest = "=0.12.22"
 serde = { workspace = true }
-serde_json = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-uuid = { version = "=1.17.0", features = ["v4", "serde"] }
+uuid = { version = "=1.17.0", features = ["serde", "v4"] }
 
-[dev-dependencies]
-mockito = { workspace = true }
+[build-dependencies]
+prost-build = { version = "=0.13.5" }
+
+[lints]
+workspace = true
