+++ crates/protocol/src/crypto.rs
