+++ crates/protocol/src/error.rs
