+++ crates/protocol/src/namespace_info.rs
